requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
instagrapi==2.0.0
Pillow==10.0.1
pycryptodomex==3.18.0
pydantic==1.10.9
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
typing_extensions==4.8.0
urllib3==2.0.5
